requirements-notorch.txt
black
blessed
numpy
pylint
pytest
pytest-cov
pytest-html
pyyaml
rich
scikit-learn
seaborn
soundfile
torchinfo
torchtyping
tqdm
uniplot
wandb
wget
